import timekeeping instead of multiple functions
